backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.9.0
gensim==4.3.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mwparserfromhell==0.6.6
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
openai==1.57.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
pyarrow==18.0.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.3.1
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
